Implemented PWM triggered ADC with DMA and interrupts. Moved PWM test to ADC interrupt where the main work will eventually happen. Started CAN bus code. Added a bunch more configuration options to EEPROM (no way to set them yet...)

+++ adc.h
@@ -1,9 +1,9 @@
 #ifndef ADC_H_
 #define ADC_H_
 void setup_adc();
-float getBusVoltage();
-float getCurrent1();
-float getCurrent2();
-float getInvTemp1();
-float getInvTemp2();
+int32_t getBusVoltage();
+int32_t getCurrent1();
+int32_t getCurrent2();
+int32_t getInvTemp1();
+int32_t getInvTemp2();
 #endif
